// src/containers/auth/type.ts
export interface IOptionLogin {
  labelInputUserEmail: string;
  labelInputPassword: string;
  labelInputButtonSubmit: string;
  labelTitle: string;
  labelForgotPassword: string;
  labelDoNotAccount: string;
  labelYourWebsite: string;
  labelInputRemember: string;
}
export interface IFormData {
  email: string;
  password: string;
  remember: boolean;
}